requirements.txt
Adafruit-BBIO==1.2.0
Adafruit-Blinka==6.15.0
adafruit-circuitpython-busdevice==5.1.1
adafruit-circuitpython-framebuf==1.4.8
adafruit-circuitpython-mcp230xx==2.5.1
adafruit-circuitpython-pct2075==1.1.11
Adafruit-PlatformDetect==3.17.2
Adafruit-PureIO==1.1.9
asyncio-mqtt==0.10.0
cbor2==5.4.2
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorlog==6.5.0
colorzero==2.0
gpio==0.3.0
idna==3.3
importlib-metadata==4.8.1
luma==2.1.3
luma.core==2.3.1
luma.oled==3.8.1
numpy==1.21.4
paho-mqtt==1.6.1
Pillow==8.4.0
pkg-resources==0.0.0
psutil==5.8.0
pyaml==21.10.1
pycparser==2.18
pyftdi==0.53.3
pyparsing==2.2.0
PyQRCode==1.2.1
pyserial==3.5
pyusb==1.2.1
PyYAML==6.0
recordclass==0.16.3
requests==2.26.0
six==1.16.0
smbus2==0.4.1
spidev==3.5
sysv-ipc==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
wrapt==1.13.3
zipp==3.6.0
